refactor(runtime-install): split install responsibilities
